docs: rename `app_legacy`
