--- docs/docs/building-apps/01-app-legacy-go.md
@@ -1,14 +1,0 @@
----
-sidebar_position: 1
----
-
-# Overview of `app_legacy.go`
-
-This section is intended to provide an overview of the `SimApp` `app_legacy.go` (or `app.go` pre-`v0.47.0`) file and is still a work in progress.
-For now please instead read the [tutorials](https://tutorials.cosmos.network) for a deep dive on how to build a chain.
-
-## Complete `app_legacy.go`
-
-```go reference
-https://github.com/cosmos/cosmos-sdk/blob/0d8787c/simapp/app_legacy.go#L103-L716
-```
